2_domashno_vot/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.6
minio==7.2.0
python-jose[cryptography]==3.3.0
requests==2.31.0
pydantic==2.6.1
pydantic-settings==2.1.0
python-dotenv==1.0.0
PyJWT==2.8.0